requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==4.2.20
django-crispy-forms==2.4
django-filter==25.1
django-htmx==1.23.0
django-polymorphic==3.1.0
django-tables2==2.7.5
et_xmlfile==2.0.0
gunicorn==23.0.0
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
setuptools>=65.0.0
django_tables2>=2.7.5
django-extensions>=4.1
django-debug-toolbar>=5.0.0
pre-commit
